chore(components): restructure the component folder
