+++ components/game/board.tsx
