+++ components/game/player-info.tsx
