+++ components/ui/accordion-item/index.ts
@@ -1,0 +1,3 @@
+import AccordionItem from "./accordion-item";
+
+export default AccordionItem;
