+++ components/ui/button/index.ts
@@ -1,0 +1,3 @@
+import Button from "./button";
+
+export default Button;
