+++ components/ui/card/index.ts
@@ -1,0 +1,3 @@
+import Card from "./card";
+
+export default Card;
