+++ components/ui/dialog/index.ts
@@ -1,0 +1,3 @@
+import Dialog from "./dialog";
+
+export default Dialog;
