+++ components/ui/input/index.ts
@@ -1,0 +1,3 @@
+import Input from "./input";
+
+export default Input;
